feat(core): split up metrics
